requirements.txt
nltk
pdfplumber
python-docx
